+++ stacks-signer/src/chainstate.rs
@@ -427,7 +427,7 @@ impl SortitionsView {
     ) -> Result<bool, SignerChainstateError> {
         // if the parent tenure is the last sortition, it is a valid choice.
         // if the parent tenure is a reorg, then all of the reorged sortitions
-        //  must either have produced zero blocks _or_ produced their first block
+        //  must either have produced zero blocks _or_ produced their first (and only) block
         //  very close to the burn block transition.
         if sortition_state.prior_sortition == sortition_state.parent_tenure_id {
             return Ok(true);
@@ -462,6 +462,23 @@ impl SortitionsView {
             if tenure.consensus_hash == sortition_state.parent_tenure_id {
                 // this was a built-upon tenure, no need to check this tenure as part of the reorg.
                 continue;
+            }
+
+            // disallow reorg if more than one block has already been signed
+            let globally_accepted_blocks =
+                signer_db.get_globally_accepted_block_count_in_tenure(&tenure.consensus_hash)?;
+            if globally_accepted_blocks > 1 {
+                warn!(
+                    "Miner is not building off of most recent tenure, but a tenure they attempted to reorg has already more than one globally accepted block.";
+                    "proposed_block_consensus_hash" => %block.header.consensus_hash,
+                    "proposed_block_signer_sighash" => %block.header.signer_signature_hash(),
+                    "parent_tenure" => %sortition_state.parent_tenure_id,
+                    "last_sortition" => %sortition_state.prior_sortition,
+                    "violating_tenure_id" => %tenure.consensus_hash,
+                    "violating_tenure_first_block_id" => ?tenure.first_block_mined,
+                    "globally_accepted_blocks" => globally_accepted_blocks,
+                );
+                return Ok(false);
             }
 
             if tenure.first_block_mined.is_some() {